pyproject.toml
[project]
name = "toy-script"
version = "0.1.0"
description = ""
authors = [{ name = "Ryan Partridge", email = "rpartridge101@gmail.com" }]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "pydantic-ai-slim[logfire] (>=0.0.19,<0.0.20)",
    "fastapi (>=0.115.6,<0.116.0)",
    "selenium (>=4.27.1,<5.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "transformers (>=4.48.0,<5.0.0)",
    "sentence-transformers (>=3.3.1,<4.0.0)",
    "logfire[asyncpg] (>=3.2.0,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu121"
priority = "explicit"


[tool.poetry.group.testing.dependencies]
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
black = "^24.10.0"
flake8 = "^7.1.1"
isort = "^5.13.2"
mypy = "^1.14.1"
flake8-docstrings = "^1.7.0"
flake8-bugbear = "^24.12.12"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"


[tool.poetry.group.ci.dependencies]
git-cliff = "^2.7.0"

[tool.pytest.ini_options]
addopts = "--cov-report term-missing --cov=tool_assistant tests/"

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
strict = true
cache_fine_grained = true
plugins = ['numpy.typing.mypy_plugin', 'pydantic.mypy']

[tool.isort]
profile = "black"

[tool.black]
line-length = 88
target-version = ['py312']
